backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
pymupdf
python-multipart
python-dotenv
sentence-transformers
faiss-cpu
spacy
reportlab
arabic-reshaper
python-bidi
